Basic App: Load & download .txt, present a snow-girl img. JS and PY, no ML.
Created a VS solution with 2 projects: Javascript/nodeJS app and Python.
The App is basic with no styles, allowing loading and downloading of 1 .txt file, presenting a fixed preview image (snow-girl) and storing the generated files in a unique directory for each call.
The app also spawns 2 Python programs: First gets the .txt file and returns a simple text via sockets. The second gets as arguments a directory and saves in it the fixed image.
There's no ML yet and no proper comments. Error handling needs to be improved.